chore: 스타일 컴포넌트 이동 및 모달 삭제
